Remove infinite recursion in `Vot::Parser`
The `Vot::Parser` class consumes a VoT or ACR value string and parses it into a set of requirements to satisfy the vector.

Some vectors have dependencies on other vectors. For example, the biometric proofing vector (`Pb`) depends on the identity proofing vector (`P1`). In the [vector component value definitions](https://github.com/18F/identity-idp/blob/5442bbdb9af5421dcb4fabc7a8ec4ce9ab307526/app/services/vot/supported_component_values.rb#L4) these are described as “implied components” i.e. vector P1 implies vector Pb.

When a partner requests a vector with a component that implies another component we expand the vector to include the implied component. For example, a partner may request `C1.Pb` which we will expand to `C1.C2.P1.Pb` since `Pb` implies `P1` and `P1` implies `C2`.

Prior to this commit these vectors depended on a recursive function which was subject to infinite recursion if vectors had mutual implications. For example, vector `A1` may imply vector `B1` which could imply vector `A1` resulting in infinite recursion expanding the vector. This commit refactors the parser to move the expansion logic into a new `Vot::ComponentExpander` which has a new implementation that does not have the infinite recursion issue.

Finally, the `implied_component_values` were constants prior to this commit. This commit makes those values strings that represent the names of the components. This was done because mutual component implication is not possible since both vectors are not defined when the first vector is defined. In this way the first vector cannot depend on the second when the second does not exist. Using strings allows the vectors to be looked up in a map after all of the vectors have been defined.

[skip changelog]