LG-13434 Link to overview article from IdV welcome
Prior to this commit the welcome screen at the begining of IdV links to a page explaining how to verify you identity. The overview page on the static site is the more appropriate page for people trying to learn more about identity verification. This commit updates the URL to link to the correct place.

[skip changelog]