LG-13434 Link to overview article from IdV welcome (#10707)
Prior to this commit the welcome screen and how to verify screen at the beginning of IdV link to a page explaining how to verify you identity. The overview page on the static site is the more appropriate page for people trying to learn more about identity verification. This commit updates the URLs to link to the correct place.

[skip changelog]