Add an unused service provider issuer argument to the resolution proofing job
We are planning on passing the SP into the background job and using it for work in the job such as computing UUID prefixes and SP costs. First we need all of the workers to know about the new argument so we do not see `ArgumentError`s when the argument is passed to the job. This commit adds the argument

[skip changelog]