LG-13383 Pass SP issuer to the ResolutionProofingJob
In #10722 we added an argument for passing the service provider issuer to the `ResolutionProofingJob` evenutally this will be used for things like computing SP costs, creating `DocAuthLog` records, and determining the UUID prefix in the job.

This commit starts passing the issuer to the job but does not start using it. In order to ensure that results are consistent a future commit will start reading the issue once all of the web hosts are passing it to the job.

This can only be safely merged once #10722 is fully deployed.

[skip changelog]