Stop passing UUID and UUID prefix to `Idv::Agent`
In #10728 we started passing the service provider ID to the resolution proofing job. Since the job has the user ID this allows us to compute UUID and UUID prefix in the job instead of having to pass them through the applicant. Once that commit is deployed we can stop passing the UUID and UUID prefix via the applicant.

This commit removes the code that looks up the UUID prefix and UUID and merges them into the applicant. This should not be merged until #10728 is fully deployed to production.

[skip changelog]