Rename the `Pb` vector to be the IAL2 conformant vector
Initially the `Pb` vector was intended to be the "biometric comparison" vector with other vectors building out the IAL2 conformant features set. Per [this slack thread](https://gsa-tts.slack.com/archives/C07152J9ACT/p1717094341824729) that is no longer the direction we are pursuing. This vector will describe IAL2 conformance and requirements will be added to the requirement set.

[skip changelog]