LG-13386 Add SP costs in the `ProgressiveProofer` class (#10767)
We are in the process of moving the logic to track SP costs from the IdP into workers. This will let us more accurately track costs by marking costly events where they actually happen (in the worker). It will also give us some flexibility with the result object that is used during resolution since it will not have SP cost tracking as a dependency.

In #10743 we stopped writing the costs in the controller if the result hash indicates that costs were already written. This commit does the work of writing costs in the `ProgressiveProofer` and returns that they were written in the result.

The next step will be removing the logic to write costs in the controller

This commit can only be safely deployed when the changes in #10743 are fully deployed.

[skip changelog]