Remove the `sp_costs_added` flag from the resolution result
The `VerifyInfoConcern` used to add SP costs to the database by inspecting the result from the `ResolutionProofingJob`. We have made the following changes in an effort to move the logic for tracking SP costs into the resolution proofing job:

- In #10743 we added a conditional to the `VerifyInfoConcern` to stop writing SP costs if a `sp_costs_added` flag is present on the result
- In #10767 we started adding costs in the `ResolutionProofingJob` and had the job set the `sp_costs_added` flag
- In #10786 we removed the code that added the SP costs in the `VerifyInfoConcern`

This commit includes the changes for the final step: Removing the `sp_costs_added` flag now that nothing is reading it.

This commit can be safely merged into main when all of the above changes are fully deployed.

[skip changelog]