Update the cloudwatch query for identity verification report to filter fraud-pending in-person users
When the identity verification report was built the fraud-review functionality did not apply to in-person pending. This was changed in #9928, but the report was not updated to filter fraud-pending users out of the successfully verified count for in-person events.

This commit modifies the cloudwatch query to filter those users who are not verified from being counted as successfully verified.

[skip changelog]