Do not track SP costs in AAMVA exception case (#10793)
I was looking at the changes in #10786 for the move of SP cost tracking from the `VerifyInfoConcern` to the `ProgressiveProofer` when I noticed this condition that was not accounted for:

https://github.com/18F/identity-idp/pull/10786/files#diff-5d8b1370b1c579364bb79a0813ca9bf9a15c65e7a974829886ab3e00f3441cedL310

This logic prevents a AAMVA cost from being added if an AAMVA exception occurs. Presumably this is because we are not billed for AAMVA exception which happen frequently. This commit applies that logic to the `ProgressiveProofer` where SP costs are now tracked.

[skip changelog]