Schedule weekly reports to run Monday for the previous week
Our jobs use the default Rails definition of a week starting on a Monday. I think this is the ISO standard. Running a job on Monday and giving it the argument `Time.zone.yesterday` means that weekly jobs will process data for the previous week

changelog: Internal, Reporting, Move weekly reports to run at 00:00 Monday UTC