Determine whether an AAMVA check is required in the progressive proofer (#10796)
Prior to this commit we have been passing `should_proof_state_id` from the `VerifyInfoConcern` down to the `ProgressiveProofer` through the async proofing tooling. This requires this argument to be passed down the call stack and stored as a background job argument.

The `should_proof_state_id` value is determined by configuration that is available to the IdP and the workers. It is not clear why it needs to be computed in the `VerifyInfoConcern` and passed down through the job to the `ProgressiveProofer`. This commit shortcuts everything by moving the logic closer to where the value is actually inspected.

[skip changelog]