LG-13301 Add more categories to the workflow completed breakdown (#10828)
The workflow completed section of the identity verification report describes the state a user is in when completing the proofing workflow. That has the dimensions:

- Fraud review pending: The user may or may not be fraud review pending
- GPO pending: The user may or may not be waiting on a letter
- In-person pending: The user may or may not need to proof in person

A user can have all, none, or some combination of these when completing the workflow. Previously the workflow steps did not account for all of the possible cases. This commit updates the report to lay out the counts for each possible case.

[skip changelog]