Add a test that shows the issue with the GPO upgrade path
I discovered an odd before action in the GPO flow. It appears that we do not let users with a pending GPO profile enter proofing. This is _fine_ except for users should be able to enter proofing with a pending GPO profile _if_ they are doing a biometric upgrade.

This commit introduces a failing spec to demostrate the issue.