Add a `Idv::ByMail::ResendLetterController` and associated routes (#10864)
This commit begins the process of splitting the `Idv::ByMail::RequestLetter` controller into 2 separate controllers:

1. A controller for requesting a letter in the IdV workflow
2. A controller for requesting a new letter after the IdV workflow from the enter-code step

This change is being made because the current `Idv::ByMail::RequestLetter` controller is built like an IdV-Step controller but has some hacks to enable it to be visited outside the IdV flow to support resends. These exceptions lead to a number of inconsistencies that risk becoming bugs as we make changes to the GPO flow.

This commit makes the first part of the change: Adding the new controller and setting up routes for it. A future change will start directing traffic to the new routes and a further change will remove the old routes.

[skip changelog]