LG-13692 Use `PendingProfileConcern` to check for pending profiles during IdV
The `PendingProfileConcern` uses the `PendingProfilePolicy` to check if a user has a pending profile and needs to be redirected to finish proofing. We have code in the `IdvStepConcern` that does the same, but does not factor all of the things that go into that function, such as whether the SP requested a biometric comparison and the pending profile had a biometric comparison.

This commit uses the `PendingProfileConcern` in the `IdvStepConcern` to ensure the pending profile logic is consistent.

[skip changelog]