Add a test to reproduce the issue in LG-13472
The issue found in LG-13472 appears to occur when a user submits the verify info step for the doc auth flow with a pending in-person enrollment. This appears to happen when the user has a barcode error and opts to go into in-person proofing. This commit does nothing to fix the bug but does add a test to identify it.

[skip changelog]