LG-13420 Add the `two_pieces_of_fair_evidence` requirement (#10871)
When a service provider requests a biometric comparison we want to make this user proof through a process that requires 2 pieces of fair evidence. This commit starts the work for that by adding a requirement to the `Pb` component. Whenever the biometric request is made this requirement will be present on the `resolved_authn_context_result`. We can use this requirement to determine if the verify-by-mail flow is available to the user.

[skip changelog]