Begin using the new `ResendLetterController` (#10865)
The `ResendLetterController` was introduced in #10864. It was not used to support the 50/50 state when old instances will not have the route and would 404 if users were linked there.

Once the changes in #10864 are fully merged this change can be merged to start linking users to the new `ResendLetterController`. From there the functionality to support resends can be removed from the `RequestLetterController`.

[skip changelog]