Remove the resend functionality from the `RequestLetterController` (#10866)
In #10864 a new controller was introduced to handle resends. In #10865 the controller started being used. Once both of those changes are fully deployed this change can be merged to remove the resend functionality from the `RequestLetterController`.

This change will make the `RequestLetterController` behave like the rest of the IdV-step controllers. It will no longer be accessible outside of the IdV workflow.

[skip changelog]