LG-13775 Retry GPO confirmation upload when SSH errors occur
We have observed a number of SSH errors when attempting to upload the GPO file. When those have occured re-running the GPO job has resulted in a successful upload. This commit attempts to automate that by adding retry logic to confirmation uploader.

This change is limited to the SSH connection and upload. This is to tread carefully around the job's logic. It deletes the records it uploads so we need to tread lightly when performing retries.

[skip changelog]