Don't allow letter sends when 2 pieces of fair evidence are required (#10886)
In #10871 we added a requirement that 2 pieces of fair evidence are required during verification. This is active whenever a SP requests biometric comparison. This commit adds enforcement of that requirement. It uses the `GpoVerifyForm` added in #10844 to disallow sends when that requirement is in place.

changelog: User-Facing Improvements, Verify By Mail Flow, The verify by mail flow enforces the 2 pieces of fair evidence requirement which disallows verify by mail when a SP requests an authn context with the requirement.