Add a FSAFEDS specific ACR value
This commit adds a new ACR value with the following value: `http://idmanagement.gov/ns/assurance/ial/2?fsa_feds=true`.

Using this value as an "IAL" value will yield the following result:

1. Identity proofed attributes are shared for users who have already proofed
2. Users with a .gov or .mil email address or a PIV/CAC assocaited with their account are not required to proof
3. Users without a .gov or .mil email address or a PIV/CAC are required to proof if they have not already

This change is made by addiing logic to the `AuthnContextResolver` to apply this policy when this ACR value is present. This logic includes the introduction of a `allowed_fsa_feds_idv_exception_providers` which is the list of service providers allowed to use this features. This allows us to limit the use of this ACR value to just FSAFEDS.

This commit also makes changes to the SAML and OIDC interfaces to make it possible to request this value.

[skip changelog]