LG-13123 Log when `GpoConfirmationUploader` uploads `GpoConfirmations`
We have observed a number of issues with the `GpoConfirmationUploader`. We are adding more visibility to the job to improve our monitoring and alerting to deal with failures that come up.

This commit adds a line to `events.log` when confirmations get uploaded. It includes a count of the confirmations that were uploaded so we can build a metric for that.

[skip changelog]