Update the verify-by-phone rate limit screen in no-GPO case
When the GPO option is available the verify-by-phone rate limit screen contains the following:

> For security reasons, we limit the number of times you can attempt to verify your phone number online.
>
> You can:
> - Verify by mail, which takes 5 to 10 days
> - Cancel and start over again after 5 hours and 59 minutes

However, when the verify-by-mail option is not available this list only contains one bullet point. This commit removes the list when there is no verify by mail option. It introduces a paragraph to tell the user to try again after the rate limit expires.

[skip changelog]