LG-13959 Fix whitespace issue with subsetted Public Sans
We subset our fonts so they only contain the glyphs we need for the content on our site. We use `glyphhanger` to remove all of the characters that are not present in an allowlist at `app/assets/fonts/glphys.txt`.

We discovered a quirk where this tool appears to whitespace on the context of this file when it reads it through the `subset` option. This is an issue since the first character in our list is whitespace. This commit adjusts the documented command to use the `whitelist` which takes a list of characters.

This commit includes the regenerated fonts.

[skip changelog]