LG-14095: set up A/B testing for Socure
This sets up a new A/B test to allow management of the distribution of
doc_auth requests between LexisNexis TrueID and Socure.

There is an already existing A/B test for swapping out doc_auth vendors,
DOC_AUTH_VENDOR, but that tests parameters are integrated with the
current code base and is really set up for TrueID vs Acuant, two very
similar vendors. Given that the Socure integration will be significantly
different, I chose to start a new test, and expect that in time it will
replace the preceding one.

[skip changelog]