LG-14253 Track time to complete proofing workflow (#11179)
We want to start tracking how long it takes people to complete identity proofing. Part of that is understanding how long it takes people to go from the start of the process to the enter password step where their profile is created. At this point we want to know:

1. How long did it take the person to complete the workflow
2. What steps remain before person finishes proofing (i.e. verify-by-mail, in-person-proofing, or fraud review)

The enter password submit events already logs enough information to determine the second. This commit starts tracking the duration of the proofing session by adding a timestamp at the welcome step and logging the elapsed seconds since on the enter password events.