LG-14380 Add `profile_age_in_seconds` to out-of-band verification events (#11186)
We are working to better measure the time it takes users to complete identity proofing. In #11179 we began logging how long it takes people to complete the identity proofing workflow.

This proofing workflow time only tells a part of the story. Many users have to complete out-of-band events to finish identity proofing. We need to know how long these workflows take to complete to fully understand how long identity proofing takes.

When a user finishes the proofing workflow we create a `Profile` record. This profile represents the proofing attempt and is in a “pending” state until the user completes those events. By logging the age of the pending profile we can determine how long it has been since the user finished the proofing workflow and thus how long they have been in a pending state.

This commit opts to log the age of the profile in seconds. This will simplify the creation of the Cloudwatch dashboard to represent this information since CloudWatch does not have good tooling for dealing with timestamps.

There are 3 out-of-band processes that can be required to complete proofing:

1. Verify-by-mail code entry
2. In-person proofing results
3. Fraud review pass

These are represented by the following events:

1.  `IdV: enter verify by mail code submitted`
2. `GetUspsProofingResultsJob: Enrollment status updated`
3. `Fraud: Profile review passed`

This commit adds a `profile_age_in_seconds` property to all of these events. To measure this a `profile_age_in_seconds` method was also added to the `Profile` model.

The `profile_age_in_seconds` property was also added to the `Fraud: Profile review rejected` events to maintain existing symmetry between arguments there.