LG-14308 Add an account verified but not connected template
This commit adds a new `account_verified_but_not_connected` email to the user mailer. This email is intended to be sent to users who verify their identity but do not continue to the service provider. The content is intended to give them a link to the service provider so they can sign in there and connect their account.

This commit also includes some tweaks to the existing `account_verified` email.

[skip changelog]