LG-14432 Log `fraud_pending_reason` on `idv_final` event
Users are marking `fraud_review_pending` on the last step of the IdV process. For users who go through document capture and verify by phone that means that they are marked `fraud_review_pending` at enter password. However, users doing verify-by-mail or in-person-pending are not marked `fraud_review_pending` until they complete those flows. This is done so that users do not get marked `fraud_review_pending` if we have disabled device profiling by the time they get to those steps.

This is an issue with logging at the enter-password step since those users in in-person and verify-by-mail that will eventually be marked fraud review pending are not when we log those properties. This is especially problematic since we use the `idv_final` event here for the Identity Verification Report.

If a user has a device profiling result that will lead them to be fraud-review pending we store that on the profile as `fraud_pending_reason` and use it downstream to move verify-by-mail and in-person users into the fraud review flow.

This commit logs the `fraud_pending_reason` so that we have visibility into users the will be eventually marked `fraud_review_pending`. It also uses this new property in the Identity Verification Report.

[skip changelog]