DO NOT MERGE: Add in-person proofing successes to the weekly drop-off report (#11075) * Add in-person proofing successes to the weekly drop-off report
changelog: User-Facing Improvements, Reporting, IdV, Add in-person proofing completion count to the weekly report that lists IdV completion statistics

* Update DropOffReport IPP queries based on PR feedback

changelog: Internal, Reporting, IdV, Consolidate query for in-person proofing statistics

<!-- Uncomment and update the sections you need for your PR! -->

<!--
## 🎫 Ticket

Link to the relevant ticket:
[LG-XXXXX](https://cm-jira.usa.gov/browse/LG-XXXXX)
-->

<!--
## 🛠 Summary of changes

Write a brief description of what you changed.
-->

<!--
## 📜 Testing Plan

Provide a checklist of steps to confirm the changes.

- [ ] Step 1
- [ ] Step 2
- [ ] Step 3
-->

<!--
## 👀 Screenshots

If relevant, include a screenshot or screen capture of the changes.

<details>
<summary>Before:</summary>

</details>

<details>
<summary>After:</summary>

</details>
-->