LG-14609 Move Threatmetrix response body to its own event
We use Threatmetrix for device profiling. When a user's device profiling transaction represents a review or reject they need to undergo the fraud review process. Part of that process is looking at what was returned by Threatmetrix in the respnose body.

The Threatmetrix response body is incredibly large. It causes some issues with Cloudwatch parsing the JSON written to the log.

This commit addresses the above by moving the Threatmetrix response into its own event. This change will allow us to continue inspecting the response body for fraud review but remove the issues with JSON parsing on the main proofing result event.

[skip changelog]