LG-13462 Move `account_lex_id` and `session_id` into `DdpResult` (#11249)
We use the `Ddp::Proofer` to communicate with Threatmetrix for device profiling. This returns a `DdpResult` which includes a large hash representing the redacted Threatmetrix response body. This result is eventually logged in the `IdV: doc auth verify proofing results` event.

The Threatmetrix response body is large and we are working to move it into its own analytics event.

The `account_lex_id` and `session_id` have been identified as especially helpful fields from the response body. Before we move the response body into its own event we are moving these fields onto `DdpResult`. This will cause them to be logged a level above the response body so they are present after we move the response body to its own event.

[skip changelog]