LG-14370 Add biographical info to proofing result (#11258)
Currently we log PII or redacted PII related to the resolution proofing job in the `IdV: doc auth verify proofing results` event. This primarily includes state ID information which is added to the logs in `VerifyInfoConcern#idv_result_to_form_response`.

This commit adds this information to the result that is returned from `ResultAdjudicator`. This has a few advantages:

1. We have a consistent result since the information is added before bing written to Redis. This will allows the `SocureShadowModeProofingJob` to pick it up and log it.
2. It provides an overview of what PII and redacted PII is logged in one place.
3. It ensures that PII is logged consistently even if vendors or stages change in the future

Finally, this commit adds the year of birth to the logs to address LG-14371.

[skip changelog]