Add NewRelic method tracers to the AbTest bucket tooling
We are considering using a different hash algorithm for the `AbTest#bucket` method.  We expect a different hashing algorithm to give us better performance which could be a meaningful bump considering how much we call the `#bucket` method every request.

This commit lays some groundwork by adding method tracers to the impacted methods. This will allow us to assess the impact this change could have.

[skip changelog]