LG-13462 Log the Threatmetrix response body in its own event (#11290)
We use Threatmetrix for device profiling. When a user's device profiling transaction represents a review or reject status they need to undergo the fraud review process. Part of that process is looking at what was returned by Threatmetrix in the logged response body.

The Threatmetrix response body is incredibly large. It causes some issues with Cloudwatch parsing the JSON written to the log.

This commit adds a new event to capture the Threatmetrix response body. The plan is to verify that this new event works for the fraud analysis purposes and then stop logging it on the existing event. The change to stop logging the response body on the existing event will be in a follow-up change.

[skip changelog]